+++ commitguard/__init__.py
@@ -2,6 +2,6 @@
 
 from .errors import AnalysisError
 
-__version__ = "0.1.3"
+__version__ = "0.2.0"
 
 __all__ = ["__version__", "AnalysisError"]